replace file-resourcing with exec for suid bit removal
since the whole file_exists attempt ate resources and kept erroring on, i took a step back and implemented a simple exec which supports a nice onlyif clause. This is not a good implementation, I really don't like it. But unlike the other one, it works as it should. So it goes in for now.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>