set LOG_OK_LOGINS = "no", correct ENV_PATH syntax
LOG_OK_LOGINS is set to "no" by default in Debian which is the upstream source
of this template. All values should be default unless they can be configured by the module.

Add separator for ENV_PATH otherwise the first value in "additional_user_paths" will not be
separated from the default value in the file.